Callback Map

+++ Util/Socket/MutexCount.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "NS.h"
 #include "../Threading/CriticalSection.h"
+
 NS_SOCKET_BEGIN
 class MutexCount
 {
@@ -8,20 +9,20 @@ public:
 	MutexCount() {}
 	virtual ~MutexCount() {}
 private:
-	unsigned long long _count;
+	unsigned long _count;
 	Util::Threading::CriticalSection _cs;
 public:
-	unsigned long long Add();
-	unsigned long long Read();
+	unsigned long Add();
+	unsigned long Read();
 };
-inline unsigned long long MutexCount::Add()
+inline unsigned long MutexCount::Add()
 {
 	_cs.EnterCriticalSection();
 	_count++;
 	_cs.LeaveCriticalSection();
 	return _count;
 }
-inline unsigned long long MutexCount::Read()
+inline unsigned long MutexCount::Read()
 {
 	return _count;
 }
